Sensor data is now sent directly to an InfluxDB
Why write another time series REST service when
there are databases like InfluxDB that exactly
do what we need.

The sketch was migrated to write to the InfluxDB.
In addition, a sensor now has a readable ID, which
allows us to distinguish individual datasets in the
InfluxDB.

Serial communication for debugging purposes is now
disabled by default. You need to uncomment the
corresponding preprocessor token in the Settings.h
to enable it.