Added a pressure sensor & static IPs
In an effort to reduce power consumption by minimizing
on times, I switched from DHCP to static IP configuration.
It saves a little time, but still not enough to realize
decent battery life (I run a sensor node on a single 18650
lithium ion cell). I still have to tweak the code, but this
is at least a first step.

I also added a BMP085 barometric pressure sensor to the
setup. With that, a single node can determine all relevant
environmental information.