Improved battery life with better connection handling
Turns out the key to reducing the on time of the sketch
was to not configure the ESPs network settings when they
were already configured. This configuration process
accounted for 50-60% of the total runtime. Now, the ESP
is only configured when it's SSID differs from the
configured one. In addition, disconnecting from the
network is not really neccessary and thus was removed.

A single activity cycle now takes only 3-5 seconds.