Optional pressure readings, better battery voltage
When deployed in a homestead, it doesn't make sense to equip
each node with a pressure sensor because the air pressure will
not vary in such a small area. To accomodate nodes without
pressure sensors, the sensor can be enabled or disabled via the
settings file.

The battery voltage is now read several times and the average value
is used. This should smoothen 'jumpy' voltage readings.