Added support for the BME280 sensor
You can now select which sensors to use by commenting/uncommenting
their definition in the settings. You cannot enable more than one
sensor for a given indicator, e.g. you can either use the HDC1000
or the BME280 to measure temperature and humidity but not both.

Also fixed the static IP address assignment. When reconnecting,
WiFi is unconfigured in terms of IP, so the proper configuration
to bypass DHCP has to run always.