Various improvements (see comment for details)
IP configuration is now done during every run, otherwise DHCP
would be used for each run after the initial. That also means
that the static configuration was not working until now. D'oh!

Allowed enabling/disabling of the PIR sensor. Until I have found
a pratical solution to suppress the large amount of RF-induced
false triggerings, I at least want to be able to switch that
thing of.

Runtime + deep sleep time now add up to a minute exactly. The
time for each run cycle is measured and the neccessary sleep time
to achieve 1 minute in total is calculated from that.

Replaces a verbatim Serial.print() with DEBUG_PRINT().